Apply Rector auto-fixes to application code
Rector prepared sets: deadCode, codeQuality, typeDeclarations (PHP 8.5)

Key changes across app/:
- Add return types to closures and arrow functions
- Replace empty() with strict comparisons (=== [])
- Explicit bool comparisons (instanceof checks instead of truthiness)
- Remove redundant float/int casts
- Simplify negated ternaries and combined ifs
- Add strict array param types from usage inference

+++ app/Http/Controllers/Api/DashboardPMActivityController.php
@@ -59,7 +59,7 @@ class DashboardPMActivityController extends Controller
                 ->pluck('turn_count');
 
             if ($turnCounts->isNotEmpty()) {
-                $avgTurnsPerPrd = (float) round($turnCounts->avg() ?? 0, 1);
+                $avgTurnsPerPrd = round($turnCounts->avg() ?? 0, 1);
             }
         }
